Consolidated plugin source code files into src directory
