+++ reach_ros/CMakeLists.txt
@@ -15,10 +15,7 @@ find_package(
              visualization_msgs)
 
 catkin_package(
-  INCLUDE_DIRS
-  include
   LIBRARIES
-  ${PROJECT_NAME}_utils
   ${PROJECT_NAME}_plugins
   CATKIN_DEPENDS
   eigen_conversions
@@ -28,19 +25,15 @@ catkin_package(
   moveit_ros_planning_interface
   visualization_msgs)
 
-# ######################################################################################################################
-# BUILD ##
-# ######################################################################################################################
+###########
+## BUILD ##
+###########
 
-include_directories(include ${catkin_INCLUDE_DIRS})
-
-# Utils Library
-add_library(${PROJECT_NAME}_utils src/utils.cpp)
-add_dependencies(${PROJECT_NAME}_utils ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-target_link_libraries(${PROJECT_NAME}_utils ${catkin_LIBRARIES} reach::reach_core)
+include_directories(${catkin_INCLUDE_DIRS})
 
 # Plugins
 add_library(${PROJECT_NAME}_plugins
+  src/utils.cpp
   # Evaluator
   src/evaluation/manipulability_moveit.cpp
   src/evaluation/joint_penalty_moveit.cpp
@@ -50,7 +43,7 @@ add_library(${PROJECT_NAME}_plugins
   # Display
   src/display/moveit_reach_display.cpp
 )
-target_link_libraries(${PROJECT_NAME}_plugins ${catkin_LIBRARIES} ${PROJECT_NAME}_utils reach::reach_core)
+target_link_libraries(${PROJECT_NAME}_plugins ${catkin_LIBRARIES} reach::reach_core)
 
 # Reach study node
 add_executable(${PROJECT_NAME}_node src/reach_study_node.cpp)
@@ -59,16 +52,13 @@ target_link_libraries(${PROJECT_NAME}_node ${catkin_LIBRARIES} yaml-cpp reach::r
 # Demo
 add_subdirectory(demo)
 
-# ######################################################################################################################
+############
 # INSTALL ##
-# ######################################################################################################################
+############
 
 install(
-  TARGETS ${PROJECT_NAME}_utils
-          ${PROJECT_NAME}_plugins
+  TARGETS ${PROJECT_NAME}_plugins
           ${PROJECT_NAME}_node
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
-
-install(DIRECTORY include/${PROJECT_NAME}/ DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION})
