+++ reach_ros/src/display/moveit_reach_display.cpp
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <reach_ros/display/moveit_reach_display.h>
-#include <reach_ros/utils.h>
+#include "moveit_reach_display.h"
+#include "../utils.h"
 
 #include <moveit/common_planning_interface_objects/common_objects.h>
 #include <moveit_msgs/PlanningScene.h>
